refactor(gateway): move status handler to status_runtime module (#3216)
